minimize rscript and consolidate predict functions
