+++ src/wisecondorx/main.py
@@ -16,17 +16,9 @@ from wisecondorx.newref_control import (
     tool_newref_merge,
 )
 from wisecondorx.newref_tools import train_gender_model, get_mask
-from wisecondorx.overall_tools import gender_correct, scale_sample, Sex
-from wisecondorx.predict_output import generate_output_tables
-from wisecondorx.plotter import write_plots
-from wisecondorx.predict_tools import (
-    log_trans,
-    exec_cbs,
-    apply_blacklist,
-    predict_gender,
-    normalize,
-    get_post_processed_result,
-)
+from wisecondorx.overall_tools import gender_correct, scale_sample
+from wisecondorx.predict import wcx_predict, wcx_sex
+
 from wisecondorx.ref_qc import qc_reference
 
 from wisecondorx import __version__
@@ -227,351 +219,6 @@ def wcx_newref(
     logging.info("Finished creating reference")
 
 
-def wcx_predict(
-    infile: Path = typer.Argument(..., help=".npz input file"),
-    reference: Path = typer.Argument(
-        ..., help="Reference .npz, as previously created with newref"
-    ),
-    outid: str = typer.Argument(
-        ...,
-        help="Basename (w/o extension) of output files (paths are allowed, e.g. path/to/ID_1)",
-    ),
-    minrefbins: int = typer.Option(
-        150,
-        "--minrefbins",
-        help="Minimum amount of sensible reference bins per target bin.",
-    ),
-    maskrepeats: int = typer.Option(
-        5,
-        "--maskrepeats",
-        help="Regions with distances > mean + sd * 3 will be masked. Number of masking cycles.",
-    ),
-    alpha: Annotated[
-        float,
-        typer.Option(
-            min=0.0,
-            max=1.0,
-            help="p-value cut-off for calling a CBS breakpoint.",
-        ),
-    ] = 1e-4,
-    cbs_nperm: Annotated[
-        int,
-        typer.Option(
-            "--cbs-nperm",
-            help="Number of permutations used by CBS.",
-            min=1,
-        ),
-    ] = 10000,
-    cbs_min_width: Annotated[
-        int,
-        typer.Option(
-            "--cbs-min-width",
-            help="Minimum markers for a CBS changed segment (2-5).",
-            min=2,
-            max=5,
-        ),
-    ] = 2,
-    cbs_undo_splits: str = typer.Option(
-        "none",
-        "--cbs-undo-splits",
-        help="Undo CBS splits using one of: none, prune, sdundo.",
-    ),
-    cbs_undo_sd: Annotated[
-        float,
-        typer.Option(
-            "--cbs-undo-sd",
-            help="SD cutoff used when --cbs-undo-splits=sdundo.",
-            min=0.0,
-        ),
-    ] = 3.0,
-    cbs_smooth: bool = typer.Option(
-        False,
-        "--cbs-smooth",
-        help="Apply DNAcopy smooth.CNA before segmentation.",
-    ),
-    zscore: Annotated[
-        float,
-        typer.Option(help="z-score cut-off for aberration calling.", min=0.0),
-    ] = 5.0,
-    beta: Annotated[
-        float,
-        typer.Option(
-            min=0.0,
-            max=1.0,
-            help="When beta is given, --zscore is ignored and a ratio cut-off is used to call aberrations.",
-        ),
-    ] = None,
-    blacklist: Path = typer.Option(
-        None,
-        "--blacklist",
-        help="Blacklist that masks regions in output, structure of header-less file: chr...(/t)startpos(/t)endpos(/n)",
-    ),
-    gender: Sex = typer.Option(
-        None,
-        "--gender",
-        help="Force WisecondorX to analyze this case as a male (M) or a female (F)",
-    ),
-    ylim: str = typer.Option(
-        "def", "--ylim", help="y-axis limits for plotting. e.g. [-2,2]"
-    ),
-    bed: bool = typer.Option(
-        True,
-        "--bed",
-        help="Outputs tab-delimited .bed files, containing the most important information",
-    ),
-    plot: bool = typer.Option(False, "--plot", help="Output .png plots"),
-    cairo: bool = typer.Option(
-        True,
-        "--cairo",
-        help="Use cairo for plotting",
-    ),
-    add_plot_title: bool = typer.Option(
-        False,
-        "--add-plot-title",
-        help="Add the output name as plot title",
-    ),
-    seed: int = typer.Option(
-        42, "--seed", help="Seed for segmentation algorithm"
-    ),
-    regions: Path = typer.Option(
-        None,
-        "--regions",
-        help="bed file with regions to be marked on the output plot",
-    ),
-) -> None:
-    """
-    Find copy number aberrations.
-    """
-    logging.info("Starting CNA prediction")
-
-    args = argparse.Namespace()
-    args.infile = infile
-    args.reference = reference
-    args.outid = outid
-    args.minrefbins = minrefbins
-    args.maskrepeats = maskrepeats
-    args.alpha = alpha
-    args.cbs_nperm = cbs_nperm
-    args.cbs_min_width = cbs_min_width
-    args.cbs_undo_splits = cbs_undo_splits
-    args.cbs_undo_sd = cbs_undo_sd
-    args.cbs_smooth = cbs_smooth
-    args.zscore = zscore
-    args.beta = beta
-    args.blacklist = blacklist
-    args.gender = gender
-    args.ylim = ylim
-    args.bed = bed
-    args.plot = plot
-    args.cairo = cairo
-    args.add_plot_title = add_plot_title
-    args.seed = seed
-    args.regions = regions
-
-    if not args.bed and not args.plot:
-        logging.critical(
-            "No output format selected. "
-            "Select at least one of the supported output formats (--bed, --plot)"
-        )
-        sys.exit()
-
-    if args.zscore <= 0:
-        logging.critical(
-            "Parameter --zscore should be a strictly positive number"
-        )
-        sys.exit()
-
-    if args.beta is not None:
-        if args.beta <= 0 or args.beta > 1:
-            logging.critical(
-                "Parameter --beta should be a strictly positive number lower than or equal to 1"
-            )
-            sys.exit()
-
-    if args.alpha <= 0 or args.alpha > 1:
-        logging.critical(
-            "Parameter --alpha should be a strictly positive number lower than or equal to 1"
-        )
-        sys.exit()
-
-    if args.cbs_undo_splits not in {"none", "prune", "sdundo"}:
-        logging.critical(
-            "Parameter --cbs-undo-splits should be one of: none, prune, sdundo"
-        )
-        sys.exit()
-
-    logging.info("Importing data ...")
-    ref_file = np.load(args.reference, encoding="latin1", allow_pickle=True)
-    sample_file = np.load(args.infile, encoding="latin1", allow_pickle=True)
-
-    sample = sample_file["sample"].item()
-    n_reads = sum([sum(sample[x]) for x in sample.keys()])
-
-    sample = scale_sample(
-        sample, int(sample_file["binsize"].item()), int(ref_file["binsize"])
-    )
-
-    gender = Sex(predict_gender(sample, ref_file["trained_cutoff"]))
-    if not ref_file["is_nipt"]:
-        if args.gender:
-            gender = args.gender
-        sample = gender_correct(sample, gender)
-        ref_gender = gender.value
-    else:
-        if args.gender:
-            gender = args.gender
-        ref_gender = "F"
-
-    logging.info("Normalizing autosomes ...")
-
-    results_r, results_z, results_w, ref_sizes, m_lr, m_z = normalize(
-        args, sample, ref_file, "A"
-    )
-
-    if not ref_file["is_nipt"]:
-        if not ref_file["has_male"] and gender == Sex.MALE:
-            logging.warning(
-                "This sample is male, whilst the reference is created with fewer than 5 males. "
-                "The female gonosomal reference will be used for X predictions. Note that these might "
-                "not be accurate. If the latter is desired, create a new reference and include more "
-                "male samples."
-            )
-            ref_gender = "F"
-
-        elif not ref_file["has_female"] and gender == Sex.FEMALE:
-            logging.warning(
-                "This sample is female, whilst the reference is created with fewer than 5 females. "
-                "The male gonosomal reference will be used for XY predictions. Note that these might "
-                "not be accurate. If the latter is desired, create a new reference and include more "
-                "female samples."
-            )
-            ref_gender = "M"
-
-    logging.info("Normalizing gonosomes ...")
-
-    null_ratios_aut_per_bin = ref_file["null_ratios"]
-    null_ratios_gon_per_bin = ref_file["null_ratios.{}".format(ref_gender)][
-        len(null_ratios_aut_per_bin) :
-    ]
-
-    results_r_2, results_z_2, results_w_2, ref_sizes_2, _, _ = normalize(
-        args, sample, ref_file, ref_gender
-    )
-
-    rem_input = {
-        "args": args,
-        "wd": str(os.path.dirname(os.path.realpath(__file__))),
-        "binsize": int(ref_file["binsize"]),
-        "n_reads": n_reads,
-        "ref_gender": ref_gender,
-        "gender": gender,
-        "mask": ref_file["mask.{}".format(ref_gender)],
-        "bins_per_chr": ref_file["bins_per_chr.{}".format(ref_gender)],
-        "masked_bins_per_chr": ref_file[
-            "masked_bins_per_chr.{}".format(ref_gender)
-        ],
-        "masked_bins_per_chr_cum": ref_file[
-            "masked_bins_per_chr_cum.{}".format(ref_gender)
-        ],
-    }
-
-    del ref_file
-
-    results_r = np.append(results_r, results_r_2)
-    results_z = np.append(results_z, results_z_2) - m_z
-    results_w = np.append(
-        results_w * np.nanmean(results_w_2),
-        results_w_2 * np.nanmean(results_w),
-    )
-    results_w = results_w / np.nanmean(results_w)
-
-    if np.isnan(results_w).any() or np.isinf(results_w).any():
-        logging.warning(
-            "Non-numeric values found in weights -- reference too small. Circular binary segmentation and z-scoring will be unweighted"
-        )
-        results_w = np.ones(len(results_w))
-
-    ref_sizes = np.append(ref_sizes, ref_sizes_2)
-
-    null_ratios = np.array(
-        [x.tolist() for x in null_ratios_aut_per_bin]
-        + [x.tolist() for x in null_ratios_gon_per_bin],
-        dtype=object,
-    )
-
-    results = {
-        "results_r": results_r,
-        "results_z": results_z,
-        "results_w": results_w,
-        "results_nr": null_ratios,
-    }
-
-    for result in results.keys():
-        results[result] = get_post_processed_result(
-            args, results[result], ref_sizes, rem_input
-        )
-
-    log_trans(results, m_lr)
-
-    if args.blacklist:
-        logging.info("Applying blacklist ...")
-        apply_blacklist(rem_input, results)
-
-    logging.info("Executing circular binary segmentation ...")
-
-    results["results_c"] = exec_cbs(rem_input, results)
-
-    if args.bed:
-        logging.info("Writing tables ...")
-        generate_output_tables(rem_input, results)
-
-    if args.plot:
-        logging.info("Writing plots ...")
-        data = {
-            "ref_gender": str(rem_input["ref_gender"]),
-            "beta": str(rem_input["args"].beta),
-            "zscore": str(rem_input["args"].zscore),
-            "binsize": str(rem_input["binsize"]),
-            "n_reads": str(rem_input["n_reads"]),
-            "cairo": rem_input["args"].cairo,
-            "results_r": results["results_r"],
-            "results_w": results["results_w"],
-            "results_c": results["results_c"],
-            "ylim": str(rem_input["args"].ylim),
-            "regions": str(rem_input["args"].regions),
-            "out_dir": str("{}.plots".format(rem_input["args"].outid)),
-        }
-
-        if rem_input["args"].add_plot_title:
-            # Strip away paths from the outid if need be
-            data["plot_title"] = str(os.path.basename(rem_input["args"].outid))
-
-        write_plots(data)
-
-    logging.info("Finished prediction")
-
-
-def wcx_gender(
-    infile: Path = typer.Argument(..., help=".npz input file"),
-    reference: Path = typer.Argument(
-        ..., help="Reference .npz, as previously created with newref"
-    ),
-) -> None:
-    """
-    Returns the sex of a .npz resulting from convert, based on a Gaussian mixture model trained during the newref phase.
-    """
-
-    ref_file = np.load(reference, encoding="latin1", allow_pickle=True)
-    sample_file = np.load(infile, encoding="latin1", allow_pickle=True)
-    gender = predict_gender(
-        sample_file["sample"].item(), ref_file["trained_cutoff"]
-    )
-    if gender == "M":
-        print("male")
-    else:
-        print("female")
-
-
 app = typer.Typer(
     name="wisecondorx",
     help="WisecondorX: Copy Number Aberration detection from Whole Genome Sequencing data.",
@@ -579,7 +226,7 @@ app = typer.Typer(
 )
 app.command(name="convert")(wcx_convert)
 app.command(name="newref")(wcx_newref)
-app.command(name="gender")(wcx_gender)
+app.command(name="gender")(wcx_sex)
 app.command(name="predict")(wcx_predict)
